fix: resolve CI failures (format, analyze, test)
- Fix notes/drafts/count return type: api.json says "number" but
  server returns int. Add endpoint_override with return_type: int.
- Support primitive type overrides (int/double/String/bool) in
  endpoint_overrides.return_type (previously only void/dynamic).
- Fix fvm detection in --build mode: check fvm availability in PATH
  before using it, fall back to plain dart command.
- Run dart fix --apply and dart format for CI compliance.

+++ lib/src/misskey_notes.dart
@@ -181,8 +181,8 @@ class MisskeyNotesDrafts {
       : _apiService = apiService;
 
   /// notes/drafts/count
-  Future<double> count() async {
-    return await _apiService.post<double>("notes/drafts/count", {});
+  Future<int> count() async {
+    return await _apiService.post<int>("notes/drafts/count", {});
   }
 
   /// notes/drafts/create
